requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2.1
Pillow==11.2.1
qrcode==8.2
reportlab==4.4.1
sqlparse==0.5.3
celery==5.4.0
redis==5.0.5